Removed implicit casts
git-svn-id: http://130.66.124.6/svn/dev_pilotage/simulator_sdk/trunk/data_source@533 5574b880-4a8e-4d59-90b7-38b80912c59c

+++ data_source/unit_tests/src/TypeCoercionTest.cpp
@@ -165,7 +165,7 @@ TEST_F(TypeCoercionTest, coerce_decoerce_floats)
           Actual: -8383272448
         Expected: (float)ref.at(k++)
         Which is: -8383272304.9414673*/
-        const float ref = -8383272304.9414673;//a.random<float>();
+        const float ref = (float)-8383272304.9414673;//a.random<float>();
         coerce(v, ref);
         double d = 0;
         decoerce(v, d);
